update lmdb-sys/lmdb to branch w/Android build fix
In order to temporarily address https://bugzilla.mozilla.org/show_bug.cgi?id=1525219 (until we can land https://bugzilla.mozilla.org/show_bug.cgi?id=1490496 and then update rkv, lmdb, and lmdb-sys to their latest versions in mozilla-central), we need a branch that contains the fix for building on Android that is described by that bug.  This is that branch.  It should be deleted once we update those crates to their latest versions in mozilla-central.